Refactor display endpoints to seperate service parT
